+++ src/lib/components/Table/TableHeader.svelte
@@ -1,10 +1,12 @@
 <script lang="ts">
-  import { cn } from "$lib"
+  import { cn } from "$lib/index.js"
+  import type { AsChild } from "$lib/types.js"
   import type { HTMLAttributes } from "svelte/elements"
   import { tv } from "tailwind-variants"
 
-  type $$Props = HTMLAttributes<HTMLTableSectionElement>
+  type $$Props = HTMLAttributes<HTMLTableSectionElement> & AsChild
 
+  export let asChild: $$Props["asChild"] = false
   export { className as class }
 
   let className = ""
@@ -16,6 +18,10 @@
   })
 </script>
 
-<thead class={cn(style.base, className)} {...$$restProps}>
+{#if asChild}
   <slot />
-</thead>
+{:else}
+  <thead class={cn(style.base, className)} {...$$restProps}>
+    <slot />
+  </thead>
+{/if}
